# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ElderHelp API",
    "description": "Verified historical research answers from approved reports. Invite-code pilot; local-only conversation history.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/v2/demo/session": {
      "post": {
        "tags": [
          "pilot"
        ],
        "summary": "Session",
        "operationId": "session_v2_demo_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/capabilities": {
      "get": {
        "tags": [
          "pilot"
        ],
        "summary": "Capabilities",
        "operationId": "capabilities_v2_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v2/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Reports",
        "operationId": "reports_v2_reports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Community"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2200,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Report",
        "operationId": "report_v2_reports__report_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports": {
      "get": {
        "tags": [
          "v1 compatibility"
        ],
        "summary": "Reports V1",
        "operationId": "reports_v1_v1_reports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Community"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2200,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/{report_id}": {
      "get": {
        "tags": [
          "v1 compatibility"
        ],
        "summary": "Report V1",
        "operationId": "report_v1_v1_reports__report_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search",
        "operationId": "search_v2_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v2/answers/stream": {
      "post": {
        "tags": [
          "answers"
        ],
        "summary": "Answer Stream",
        "operationId": "answer_stream_v2_answers_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE start, progress, verified delta, complete; typed error on failure",
            "content": {
              "text/event-stream": {}
            },
            "x-sse-events": {
              "start": "StartEvent",
              "progress": "ProgressEvent",
              "delta": "DeltaEvent",
              "complete": "CompleteV2",
              "error": "ErrorEvent"
            }
          },
          "401": {
            "description": "Invite/session required"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Invalid input"
          },
          "429": {
            "description": "Quota/concurrency limit; Retry-After header"
          },
          "503": {
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/answers/stream": {
      "post": {
        "tags": [
          "v1 compatibility"
        ],
        "summary": "Answer Stream V1",
        "operationId": "answer_stream_v1_v1_answers_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE start, verified delta, complete; typed error on failure",
            "content": {
              "text/event-stream": {}
            },
            "x-sse-events": {
              "start": "StartEvent",
              "delta": "DeltaEvent",
              "complete": "AnswerComplete",
              "error": "ErrorEvent"
            }
          },
          "401": {
            "description": "Invite/session required"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Invalid input"
          },
          "429": {
            "description": "Quota/concurrency limit; Retry-After header"
          },
          "503": {
            "description": "Required dependency unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Capabilities": {
        "properties": {
          "generation_available": {
            "type": "boolean",
            "title": "Generation Available"
          },
          "search_available": {
            "type": "boolean",
            "title": "Search Available"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "corpus_generation": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Generation"
          },
          "supported_filters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Filters",
            "default": [
              "report_ids",
              "community",
              "year_from",
              "year_to"
            ]
          }
        },
        "type": "object",
        "required": [
          "generation_available",
          "search_available"
        ],
        "title": "Capabilities"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "publisher": {
            "type": "string",
            "title": "Publisher"
          },
          "community": {
            "type": "string",
            "title": "Community"
          },
          "publication_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "suggested_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Questions"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "publisher",
          "community",
          "source_url",
          "page_count"
        ],
        "title": "ReportDetail"
      },
      "ReportDetailV2": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "publisher": {
            "type": "string",
            "title": "Publisher"
          },
          "community": {
            "type": "string",
            "title": "Community"
          },
          "publication_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "suggested_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Questions"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "acquired_at": {
            "type": "string",
            "format": "date-time",
            "title": "Acquired At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "publisher",
          "community",
          "source_url",
          "page_count",
          "revision_id",
          "acquired_at"
        ],
        "title": "ReportDetailV2"
      },
      "ReportList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ReportList"
      },
      "ReportListV2": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "ReportListV2"
      },
...
      "SearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Items"
          },
          "mode": {
            "type": "string",
            "enum": [
              "keyword",
              "hybrid"
            ],
            "title": "Mode"
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items",
          "mode"
        ],
        "title": "SearchResponse"
      },
      "SessionRequest": {
        "properties": {
          "invite_code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 12,
            "title": "Invite Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite_code"
        ],
        "title": "SessionRequest"
      },
      "SessionResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "SessionResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}